refactor(ai-routers): extract orchestration and tooling support modules
Split non-HTTP concerns out of the orchestration and tools routers by introducing shared support modules for downstream auth headers, orchestration request/capability helpers, and tool-catalog models plus metadata extraction.

The routers now act more like thin adapters: orchestration_router imports its request and capability helpers from services.orchestration_support, while tools_router imports its schema and catalog logic from services.tool_catalog. This trims a large amount of inline helper code from the router files without changing the external API shape.

Validated with 'python3 -m py_compile ai-orchestration/ai-orchestration-layer/src/services/downstream_headers.py ai-orchestration/ai-orchestration-layer/src/services/orchestration_support.py ai-orchestration/ai-orchestration-layer/src/services/tool_catalog.py ai-orchestration/ai-orchestration-layer/src/routers/orchestration_router.py ai-orchestration/ai-orchestration-layer/src/routers/tools_router.py'.
